split further main and show_detections, enclose only label in box
The main function only processes program options and delegates the rest
to the function 'detect'. Separate function to draw a box on the image.
In visualization, the white bar with detection details has been replaced
by a reen box of the same size of the text.
Rename yolo_main in yolo_detect.