replace batchnorm with scale, handle relu to FC layers
In YOLO the batchnorm operator is a mean subtraction followed by scaling
and the conv bias is applied at the end; this is better implemented in
Caffe by swapping mean and bias in convolution and use a scaling layer.
The scaling factor is computed by hand from the variance and scaling
factor in YOLO.
Handle leaky RelU for fully connected layers and support transpose
weights for convolutional layers.
Conversion now working on tiny-yolo and yolo-small.