support Darknet v2 output, tweak NMS and box coordinates
If the output is a 3-dim tensor, the convention of the 'region' layer in
Darknet v2 is used. The non-maxima suppression parameters have been
changed to have results more similar to the ones reported for YOLO.
Moreover, a manual offset of half-cell is added to the boxes to fix a
systematic shift toward the bottom-right.